requirements2.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
Pillow>=10.2.0
zeroconf>=0.131.0